model_api/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
facenet-pytorch==2.6.0
fastapi==0.115.3
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
idna==3.10
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.10.0.84
pillow==10.2.0
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.3
sniffio==1.3.1
starlette==0.41.0
sympy==1.13.1
tqdm==4.66.5
typing_extensions==4.9.0
urllib3==2.2.3
uvicorn==0.32.0
